OpenAI compatible AI provider
This commit introduces the possibility to switch to an OpenAI compatible AI provider instead of Burp AI.
This is needed because

- Burp AI data is processed in the US. Often, European companies have have DPA's which categorically prevent data transfers to the US. This prevents EU companies of using Burp AI.
- Burp AI logs all requests (content + metadata) but does not define a specific retention. A properly defined (and often configurable) retention is required by DPAs for companies to enable customers to exercise their GDPR defined rights.

OpenAI compatibility enables use of e.g. managed Azure OpenAI services, which leaves the control of the AI model to the hands of the tester, and enables them to solve above issues.

The implementation is somewhat simple. The UI is a bit silly in few areas:

- OpenAI configuration is visible in the settings even though Burp AI is selected. This is difficult to fix due to the dynamic nature of building the settings dialog.
- Burp AI must be enabled even though it's not used (by the extension).